requirements.txt
discord.py==1.5.0
requests==2.24.0
discord==1.0.1
Pillow==7.2.0
PyYAML==5.3.1
